[MIG] purchase_open_qty: Migration to 18.0
Move models and views into model specific files.
