+++ purchase_open_qty/__manifest__.py
@@ -6,7 +6,7 @@
     "name": "Purchase Open Qty",
     "summary": "Allows to identify the purchase orders that have quantities "
     "pending to invoice or to receive.",
-    "version": "17.0.2.0.0",
+    "version": "18.0.1.0.0",
     "author": "ForgeFlow, Odoo Community Association (OCA)",
     "website": "https://github.com/OCA/purchase-workflow",
     "category": "Purchases",
@@ -14,7 +14,10 @@
         "purchase_stock",
         "purchase_order_line_menu",
     ],
-    "data": ["views/purchase_view.xml"],
+    "data": [
+        "views/purchase_order_line_views.xml",
+        "views/purchase_order_views.xml",
+    ],
     "pre_init_hook": "pre_init_hook",
     "license": "AGPL-3",
     "installable": True,
