+++ purchase_open_qty/models/__init__.py
@@ -1,3 +1,4 @@
 # License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl.html).
 
 from . import purchase_order
+from . import purchase_order_line
